genericized route filter
