+++ c7n/resources/glue.py
@@ -20,7 +20,7 @@ from c7n.manager import resources, ResourceManager
 from c7n.query import QueryResourceManager, TypeInfo
 from c7n.utils import local_session, chunks, type_schema
 from c7n.actions import BaseAction, ActionRegistry
-from c7n.filters.vpc import SubnetFilter, SecurityGroupFilter
+from c7n.filters.vpc import SubnetFilter, SecurityGroupFilter, RouteTableFilter
 from c7n.filters.related import RelatedResourceFilter
 from c7n.tags import universal_augment
 from c7n.filters import StateTransitionFilter, ValueFilter, FilterRegistry, CrossAccountAccessFilter
@@ -44,8 +44,9 @@ class ConnectionSubnetFilter(SubnetFilter):
 
     RelatedIdsExpression = 'PhysicalConnectionRequirements.SubnetId'
 
+
 @GlueConnection.filter_registry.register('route-table')
-class ConnectionRouteTableFilter(ValueFilter):
+class ConnectionRouteTableFilter(RouteTableFilter):
     """Filter Route Tables connected to a Glue Dev Endpoint
 
     :example:
@@ -54,40 +55,16 @@ class ConnectionRouteTableFilter(ValueFilter):
 
         policies:
           - name: glue-endpoint-public-route
-            resource: aws.glue-dev-endpoint
+            resource: aws.glue-connection
             filters:
                 - type: route-table
                   key: RouteTables[].Routes[].DestinationCidrBlock
                   value: 0.0.0.0/0
                   op: contains
     """
-    annotation_key = 'c7n:route-table'
-    annotate = False  # no annotation from value filter
-    schema = type_schema('route-table', rinherit=ValueFilter.schema)
-    schema_alias = False
-
-
-    def process(self, resources, event=None):
-        self.augment([r for r in resources if self.annotation_key not in r])
-        return super(ConnectionRouteTableFilter, self).process(resources, event)
-
-    def augment(self, resources):
-        client = local_session(self.manager.session_factory).client('ec2')
-        for r in resources:
-            try:
-                r[self.annotation_key] = client.describe_route_tables(
-                    Filters=[
-                        {
-                            "Name":"association.subnet-id",
-                            "Values": [r['PhysicalConnectionRequirements']['SubnetId']]
-                        }
-                    ])
-                print(r[self.annotation_key])
-            except client.exceptions.EntityNotFoundException:
-                r[self.annotation_key] = {}
-
-    def __call__(self, r):
-        return super(ConnectionRouteTableFilter, self).__call__(r[self.annotation_key])
+    association_type = "association.subnet-id"
+    def _get_filter(self, r):
+        return r['PhysicalConnectionRequirements']['SubnetId']
 
 
 @GlueConnection.filter_registry.register('security-group')
@@ -150,7 +127,7 @@ class EndpointSubnetFilter(SubnetFilter):
 
 
 @GlueDevEndpoint.filter_registry.register('route-table')
-class EndpointRouteTableFilter(ValueFilter):
+class EndpointRouteTableFilter(RouteTableFilter):
     """Filter Route Tables connected to a Glue Dev Endpoint
 
     :example:
@@ -166,32 +143,9 @@ class EndpointRouteTableFilter(ValueFilter):
                   value: 0.0.0.0/0
                   op: contains
     """
-    annotation_key = 'c7n:route-table'
-    annotate = False  # no annotation from value filter
-    schema = type_schema('route-table', rinherit=ValueFilter.schema)
-    schema_alias = False
-
-
-    def process(self, resources, event=None):
-        self.augment([r for r in resources if self.annotation_key not in r])
-        return super(EndpointRouteTableFilter, self).process(resources, event)
-
-    def augment(self, resources):
-        client = local_session(self.manager.session_factory).client('ec2')
-        for r in resources:
-            try:
-                r[self.annotation_key] = client.describe_route_tables(
-                    Filters=[
-                        {
-                            "Name":"association.subnet-id",
-                            "Values": [r['SubnetId']]
-                        }
-                    ])
-            except client.exceptions.EntityNotFoundException:
-                r[self.annotation_key] = {}
-
-    def __call__(self, r):
-        return super(EndpointRouteTableFilter, self).__call__(r[self.annotation_key])
+    association_type = "association.subnet-id"
+    def _get_filter(self, r):
+        return r['SubnetId']
 
 
 @GlueDevEndpoint.action_registry.register('delete')
